requirements/production.txt
trafaret_config==2.0.2
markdown2==2.3.7
aiohttp==3.6.1
aiohttp-apispec==1.5.0
aiohttp-debugtoolbar==0.5.0
aiohttp-jinja2==1.1.2
aioredis==1.3.0
apispec==3.0.0
asn1crypto==1.0.1
async-timeout==3.0.1
attrs==19.2.0
Automat==0.7.0
cffi==1.12.3
chardet==3.0.4
Click==7.0
constantly==15.1.0
cryptography==2.7
cssselect==1.1.0
hiredis==1.0.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
Jinja2==2.10.3
lxml==4.4.1
MarkupSafe==1.1.1
marshmallow==3.2.1
multidict==4.5.2
parsel==1.5.2
pika==1.1.0
pyasn1==0.4.7
pyasn1-modules==0.2.6
pycparser==2.19
PyDispatcher==2.0.5
PyHamcrest==1.9.0
pyOpenSSL==19.0.0
PyYAML==5.1.2
queuelib==1.5.0
redis==3.3.8
Scrapy==1.7.3
service-identity==18.1.0
six==1.12.0
trafaret==1.2.0
trafaret-config==2.0.2
Twisted==19.7.0
ujson==1.35
uvloop==0.13.0
w3lib==1.21.0
webargs==5.5.2
yarl==1.3.0
zope.interface==4.6.0
Sphinx==1.8.4
sphinx-autodoc-typehints==1.6.0
sphinxcontrib-websupport==1.1.0
doc8==0.8.0
